+++ src/Phan/CLI.php
@@ -1519,7 +1519,7 @@ $init_help
   Add colors to the outputted issues.
   This is recommended for only --output-mode=text (the default) and 'verbose'
 
-  [--color-scheme={default,code,light,eclipse_dark,vim}]
+  [--color-scheme={default,code,eclipse_dark,vim,light,light_high_contrast}]
     This (or the environment variable PHAN_COLOR_SCHEME) can be used to set the color scheme for emitted issues.
 
  -p, --progress-bar, --no-progress-bar, --long-progress-bar